requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
# created-by: conda 24.11.3
accelerate=1.3.0=pypi_0
aiohappyeyeballs=2.4.4=pypi_0
aiohttp=3.11.11=pypi_0
aiosignal=1.3.2=pypi_0
altair=5.5.0=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.8.0=pypi_0
async-timeout=4.0.3=pypi_0
attrs=24.3.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
blinker=1.9.0=pypi_0
blobfile=3.0.0=pypi_0
bzip2=1.0.8=h80987f9_6
ca-certificates=2024.12.31=hca03da5_0
cachetools=5.5.1=pypi_0
certifi=2024.12.14=pypi_0
charset-normalizer=3.4.1=pypi_0
click=8.1.8=pypi_0
dataclasses-json=0.6.7=pypi_0
deprecated=1.2.18=pypi_0
dirtyjson=1.0.8=pypi_0
diskcache=5.6.3=pypi_0
distro=1.9.0=pypi_0
exceptiongroup=1.2.2=pypi_0
faiss-cpu=1.9.0.post1=pypi_0
filelock=3.17.0=pypi_0
filetype=1.2.0=pypi_0
fire=0.7.0=pypi_0
frozenlist=1.5.0=pypi_0
fsspec=2024.12.0=pypi_0
gguf=0.14.0=pypi_0
gitdb=4.0.12=pypi_0
gitpython=3.1.44=pypi_0
gmp=6.2.1=hc377ac9_3
gmpy2=2.1.2=py310h8c48613_0
greenlet=3.1.1=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.7=pypi_0
httpx=0.28.1=pypi_0
httpx-sse=0.4.0=pypi_0
huggingface-hub=0.27.1=pypi_0
idna=3.10=pypi_0
jinja2=3.1.5=pypi_0
jiter=0.8.2=pypi_0
joblib=1.4.2=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=3.0.0=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2024.10.1=pypi_0
langchain=0.3.15=pypi_0
langchain-community=0.3.15=pypi_0
langchain-core=0.3.31=pypi_0
langchain-text-splitters=0.3.5=pypi_0
langsmith=0.3.1=pypi_0
libcxx=14.0.6=h848a8c0_0
libffi=3.4.4=hca03da5_1
llama-cloud=0.1.11=pypi_0
llama-cpp-python=0.3.6=pypi_0
llama-index=0.12.14=pypi_0
llama-index-agent-openai=0.4.3=pypi_0
llama-index-cli=0.4.0=pypi_0
llama-index-core=0.12.14=pypi_0
llama-index-embeddings-openai=0.3.1=pypi_0
llama-index-indices-managed-llama-cloud=0.6.4=pypi_0
llama-index-llms-openai=0.3.14=pypi_0
llama-index-multi-modal-llms-openai=0.4.2=pypi_0
llama-index-program-openai=0.3.1=pypi_0
llama-index-question-gen-openai=0.3.0=pypi_0
llama-index-readers-file=0.4.4=pypi_0
llama-index-readers-llama-parse=0.4.0=pypi_0
llama-models=0.0.63=pypi_0
llama-parse=0.5.20=pypi_0
llama-stack=0.0.63=pypi_0
llama-stack-client=0.0.63=pypi_0
llvm-openmp=14.0.6=hc6e5704_0
lxml=5.3.0=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.2=pypi_0
marshmallow=3.26.0=pypi_0
mdurl=0.1.2=pypi_0
mpc=1.1.0=h8c48613_1
mpfr=4.0.2=h695f6f0_1
mpmath=1.3.0=py310hca03da5_0
multidict=6.1.0=pypi_0
mypy-extensions=1.0.0=pypi_0
narwhals=1.23.0=pypi_0
ncurses=6.4=h313beb8_0
nest-asyncio=1.6.0=pypi_0
networkx=3.4.2=py310hca03da5_0
nltk=3.9.1=pypi_0
numpy=1.26.4=pypi_0
openai=1.60.2=pypi_0
openssl=3.0.15=h80987f9_0
orjson=3.10.15=pypi_0
packaging=24.2=pypi_0
pandas=2.2.3=pypi_0
pillow=11.1.0=pypi_0
pinecone=5.4.2=pypi_0
pinecone-plugin-inference=3.1.0=pypi_0
pinecone-plugin-interface=0.0.7=pypi_0
pip=24.2=py310hca03da5_0
prompt-toolkit=3.0.50=pypi_0
propcache=0.2.1=pypi_0
protobuf=4.25.5=pypi_0
psutil=6.1.1=pypi_0
pyaml=25.1.0=pypi_0
pyarrow=19.0.0=pypi_0
pycryptodomex=3.21.0=pypi_0
pydantic=2.10.6=pypi_0
pydantic-core=2.27.2=pypi_0
pydantic-settings=2.7.1=pypi_0
pydeck=0.9.1=pypi_0
pygments=2.19.1=pypi_0
pymupdf=1.25.2=pypi_0
pypdf=5.2.0=pypi_0
python=3.10.16=hb885b13_1
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
pytz=2024.2=pypi_0
pyyaml=6.0.2=py310h80987f9_0
readline=8.2=h1a28f6b_0
referencing=0.36.1=pypi_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
requests-toolbelt=1.0.0=pypi_0
rich=13.9.4=pypi_0
rpds-py=0.22.3=pypi_0
safetensors=0.5.2=pypi_0
scikit-learn=1.6.1=pypi_0
scipy=1.15.1=pypi_0
sentence-transformers=3.4.0=pypi_0
sentencepiece=0.2.0=pypi_0
setuptools=75.1.0=py310hca03da5_0
six=1.17.0=pypi_0
smmap=5.0.2=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.6=pypi_0
sqlalchemy=2.0.37=pypi_0
sqlite=3.45.3=h80987f9_0
streamlit=1.41.1=pypi_0
streamlit-pdf-viewer=0.0.20=pypi_0
striprtf=0.0.26=pypi_0
sympy=1.13.1=pypi_0
tenacity=9.0.0=pypi_0
termcolor=2.5.0=pypi_0
threadpoolctl=3.5.0=pypi_0
tiktoken=0.8.0=pypi_0
tk=8.6.14=h6ba3021_0
tokenizers=0.21.0=pypi_0
toml=0.10.2=pypi_0
torch=2.2.2=pypi_0
tornado=6.4.2=pypi_0
tqdm=4.67.1=pypi_0
transformers=4.48.1=pypi_0
typing-inspect=0.9.0=pypi_0
typing_extensions=4.12.2=py310hca03da5_0
tzdata=2025.1=pypi_0
urllib3=2.3.0=pypi_0
watchdog=6.0.0=pypi_0
wcwidth=0.2.13=pypi_0
wheel=0.44.0=py310hca03da5_0
wrapt=1.17.2=pypi_0
xz=5.4.6=h80987f9_1
yaml=0.2.5=h1a28f6b_0
yarl=1.18.3=pypi_0
zlib=1.2.13=h18a0788_1
zstandard=0.23.0=pypi_0
